Improving template support

+++ client/js/services/app/sortableAppService.js
@@ -68,34 +68,28 @@
         }
 
         function setOptions(originalElm, droppedElm) {
-            var wrapperOriginalRowsScope    = angular.element(originalElm.closest('.rows')).scope().$parent,
-                wrapperDropRowsScope        = angular.element(droppedElm.closest('.rows')).scope().$parent;
-            console.log("wrapping drop rows", angular.element(droppedElm.closest('.rows')).scope());
             options = {};
-            options.isNewItem           = droppedElm.attr('sortable-add-app') !== undefined;
-            options.originalRows        = portalService.getWrappingRows(wrapperOriginalRowsScope);
-            //options.dropRows            = portalService.getWrappingRows(wrapperDropRowsScope);
-            options.dropRows            = (angular.element(droppedElm.closest('.rows')).scope().row.bla)
-                ? $rootScope.portal.template.rows
-                : angular.element(droppedElm.closest('.rows')).scope().$parent.page.rows;
-            options.droppedElm          = droppedElm;
-            options.originalColIndex    = originalElm.closest('.columns').prevAll('.columns').size();
-            options.dropColIndex        = droppedElm.closest('.columns').prevAll('.columns').size();
-            options.originalRowIndex    = originalElm.closest('.rows').prevAll('.rows').size();
-            options.originalAppIndex    = originalElm.prevAll('.app').size();
-            options.dropRowIndex        = droppedElm.closest('.rows').prevAll('.rows').size();
-            options.dropAppIndex        = droppedElm.prevAll('.app').size();
-            options.originalRow         = originalElm.closest('.rows').scope().row;
-            options.dropRow             = droppedElm.closest('.rows').scope().row;
-            options.dropCol             = options.dropRow.columns[options.dropColIndex];
-            options.originalCol         = options.originalRow.columns[options.originalColIndex];
-            options.elmHasChangedRow    = options.originalRow.$$hashKey !== options.dropRow.$$hashKey;
-            options.affectedDropCol     = colService.getAffectedCol(options.dropRow.columns, options.dropColIndex);
-            options.affectedOriginalCol = (options.elmHasChangedRow)
+            options.isNewItem               = droppedElm.attr('sortable-add-app') !== undefined;
+            options.wrapperOriginalRowScope = angular.element(originalElm.closest('.rows')).scope();
+            options.wrapperDropRowScope     = angular.element(droppedElm.closest('.rows')).scope();
+            options.originalRows            = rowService.getWrappingRows(options.wrapperOriginalRowScope);
+            options.dropRows                = rowService.getWrappingRows(options.wrapperDropRowScope);
+            options.droppedElm              = droppedElm;
+            options.originalColIndex        = originalElm.closest('.columns').prevAll('.columns').size();
+            options.dropColIndex            = droppedElm.closest('.columns').prevAll('.columns').size();
+            options.originalRowIndex        = originalElm.closest('.rows').prevAll('.rows').size();
+            options.originalAppIndex        = originalElm.prevAll('.app').size();
+            options.dropRowIndex            = droppedElm.closest('.rows').prevAll('.rows').size();
+            options.dropAppIndex            = droppedElm.prevAll('.app').size();
+            options.originalRow             = originalElm.closest('.rows').scope().row;
+            options.dropRow                 = droppedElm.closest('.rows').scope().row;
+            options.dropCol                 = options.dropRow.columns[options.dropColIndex];
+            options.originalCol             = options.originalRow.columns[options.originalColIndex];
+            options.elmHasChangedRow        = options.originalRow.$$hashKey !== options.dropRow.$$hashKey;
+            options.affectedDropCol         = colService.getAffectedCol(options.dropRow.columns, options.dropColIndex);
+            options.affectedOriginalCol     = (options.elmHasChangedRow)
                 ? colService.getAffectedCol(options.originalRow.columns, options.originalColIndex)
                 : options.affectedDropCol;
-
-            console.log("DP", options.dropRows, angular.element(droppedElm.closest('.rows')).scope())
             return options;
         }
 
@@ -129,9 +123,7 @@
         }
 
         function updateWithRowChange() {
-            console.log("adding empty rows!!", options.dropRows);
-
-            if (rowService.isDropRowEmpty(options.dropRow)) {
+            if (rowService.isDropRowEmpty(options.dropRow) && !rowService.isTemplateRow(options.wrapperDropRowScope.row)) {
                 addEmptyRows();
             }
             if (options.affectedDropCol && options.dropCol.apps.length > 0) {
@@ -192,7 +184,8 @@
                 colService.addEmptyCols(options.dropRow, options.dropCol, options.dropColIndex);
             }
             colService.normalizeEmptyCols(options.originalRow);
-            if (rowService.isOriginalRowEmpty(options.originalRow)) {
+            if (rowService.isOriginalRowEmpty(options.originalRow)
+            && !rowService.isTemplateRow(options.wrapperOriginalRowScope.row)) {
                 rowService.deleteRowAndDependencies(options.originalRows, options.originalRowIndex);
             }
         }
